Iteration VII - added solution without await
